requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
twilio==9.4.1
openai==1.59.3
python-dotenv==1.0.1
aiosqlite==0.20.0
jinja2==3.1.4
httpx==0.28.1
pydantic==2.10.4
python-multipart==0.0.18
